develop -> main

+++ src/main/java/com/example/demo/domain/group/GroupServiceImpl.java
@@ -2,10 +2,9 @@ package com.example.demo.domain.group;
 
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import javax.management.InstanceAlreadyExistsException;
+import javax.management.InstanceNotFoundException;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
@@ -15,21 +14,18 @@ import java.util.UUID;
 @RequiredArgsConstructor
 @Transactional
 public class GroupServiceImpl implements GroupService {
-
-    @Autowired
     private final GroupRepository groupRepository;
 
     @Override
     public List<Group> findAll() {return groupRepository.findAll();}
 
     @Override
-    public Optional<Group> findById(UUID id) {
-        return groupRepository.findById(id);
-    }
-
-    @Override
-    public Group findMembersByGroupname(String groupname){
-        return groupRepository.findByGroupname(groupname);
+    public Optional<Group> findById(UUID id) throws InstanceNotFoundException {
+        if(groupRepository.existsById(id)) {
+            return groupRepository.findById(id);
+        } else {
+            throw new InstanceNotFoundException("Group matching {id:\"" + id + "\"} does not exist");
+        }
     }
 
     @Override
@@ -44,23 +40,21 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public void updateGroup(UUID id, Group newGroup) {
-//        Group group = groupRepository.getById(id);
-//        if(newGroup.getId().equals("")) {
-//            group.setId(id);
-//        } else {
-//            group.setId(newGroup.getId());
-//        }
-//        group.setGroupname(newGroup.getGroupname());
-//        group.setMotto(newGroup.getMotto());
-//        group.setUsers(newGroup.getUsers());
-
-        newGroup.setId(id);
-        groupRepository.save(newGroup);
+    public Group updateGroup(UUID id, Group newGroup) throws InstanceNotFoundException {
+        if(groupRepository.existsById(id)) {
+            newGroup.setId(id);
+            return groupRepository.save(newGroup);
+        } else {
+            throw new InstanceNotFoundException("Group matching {id:\"" + id + "\"} does not exist");
+        }
     }
 
     @Override
-    public void deleteGroup(UUID id) {
-        groupRepository.deleteById(id);
+    public void deleteGroup(UUID id) throws InstanceNotFoundException {
+        if(groupRepository.existsById(id)) {
+            groupRepository.deleteById(id);
+        } else {
+            throw new InstanceNotFoundException("Group matching {id:\"" + id + "\"} does not exist");
+        }
     }
 }
